+++ src/main/java/net/silentchaos512/scalinghealth/item/HeartCrystal.java
@@ -36,7 +36,9 @@ public class HeartCrystal extends StatBoosterItem {
 
     @Override
     boolean isStatIncreaseAllowed(PlayerEntity player, IPlayerData data) {
-        return Players.heartCrystalsIncreaseHealth(player)
+        //if this would increase the health by 0, we wont't process it
+        //TODO not sure if I clamp the hp. If the increase is of 5 and max hp is of 41, it could be possible to get 45 hp. (have to check)
+        return Players.heartCrystalIncreaseAmount(player) > 0
                 && data.getExtraHearts() < Players.maxHeartCrystals(player);
     }
 
@@ -59,7 +61,7 @@ public class HeartCrystal extends StatBoosterItem {
 
     @Override
     void increaseStat(PlayerEntity player, ItemStack stack, IPlayerData data) {
-        data.addHeart(player);
+        data.addHearts(player, Players.heartCrystalIncreaseAmount(player));
     }
 
     @Override