+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.9] - 2020-01-15
+### Added
+- Config for changing heart crystal heart gain
+- Config for blacklisting mods in damage scaling
+### Fixed
+- Xp client desync bug
+
 ## [2.3.8] - 2020-01-12
 ### Added
 - Functionality for configs: